+++ app/components/Base/Keypad/useCurrency.ts
@@ -2,7 +2,13 @@ import { useMemo } from 'react';
 import { CURRENCIES } from './constants';
 import createKeypadRule from './createKeypadRule';
 
-function useCurrency(currency, decimals) {
+interface UseCurrencyResult {
+  handler: (currentAmount: string, inputKey: string) => string;
+  symbol: string | null;
+  decimalSeparator: string | null;
+}
+
+function useCurrency(currency?: string, decimals?: number): UseCurrencyResult {
   const currencyData = useMemo(() => {
     if (!currency) {
       return CURRENCIES.default;
@@ -15,7 +21,7 @@ function useCurrency(currency, decimals) {
       return existingCurrency;
     }
 
-    if (decimals > 0) {
+    if (decimals && decimals > 0) {
       return {
         decimalSeparator: '.',
         handler: createKeypadRule({ decimalSeparator: '.', decimals }),